requirements.txt
colorama
Levenshtein
loguru
MaaAgentBinary
MaaFw==5.4.2
numpy
pillow
python-Levenshtein
RapidFuzz
StrEnum
requests
